Modificacion de back

+++ Cuarta-entrega/cuartaEnterega-4/src/main/java/com/example/demo/services/DetalleDeOrdenesServices.java
@@ -6,7 +6,6 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-
 import com.example.demo.models.DetalleDeLaOrden;
 import com.example.demo.models.OrdenDeCompra;
 import com.example.demo.repositories.DetalleDeLaOrdenRepository;
@@ -15,11 +14,11 @@ import com.example.demo.repositories.DetalleDeLaOrdenRepository;
 public class DetalleDeOrdenesServices {
 	@Autowired
 	DetalleDeLaOrdenRepository detalleDeLaOrdenRepository;
-	
+
 	public List<DetalleDeLaOrden> obtenerDetalleDeOrden() {
 		return detalleDeLaOrdenRepository.findAll();
 	}
-	
+
 	public Optional<DetalleDeLaOrden> findById(Integer id) {
 		return detalleDeLaOrdenRepository.findById(id);
 	}
@@ -28,8 +27,8 @@ public class DetalleDeOrdenesServices {
 		detalleDeLaOrdenRepository.save(detalleOrden);
 		return "Detalle de la orden creada correctamente";
 	}
-	
-	public String modificarDetalleOrden( DetalleDeLaOrden detalleOrden) {
+
+	public String modificarDetalleOrden(DetalleDeLaOrden detalleOrden) {
 		try {
 			DetalleDeLaOrden or = detalleDeLaOrdenRepository.findById(detalleOrden.getId()).get();
 			or.setDetalleCantidad(detalleOrden.getDetalleCantidad());
@@ -41,16 +40,16 @@ public class DetalleDeOrdenesServices {
 		}
 	}
 
-
 	public DetalleDeLaOrden eliminarDetalleDeOrden(Integer id) {
 		try {
-			DetalleDeLaOrden detalleToReturn = detalleDeLaOrdenRepository.findById(id).get();
-			detalleDeLaOrdenRepository.deleteById(id);
-			return detalleToReturn;
+			Optional<DetalleDeLaOrden> detalleToReturn = detalleDeLaOrdenRepository.findById(id);
+			if (detalleToReturn.isPresent()) {
+				detalleDeLaOrdenRepository.deleteById(id);
+			}
 		} catch (Exception err) {
 			err.getMessage();
 		}
 		return null;
 	}
-	
+
 }
